+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+# 8.001
+ * Moved to new hamf,dtype using the hamf protocols.  See hamf's defprotocol namespace for reasons.  If you are
+   extending dataset protocols via extend, extend-type or extend-protocol you need to use defprotocol's drop-in
+   replacements as opposed to clojure's default implementations.
+ 
 # 7.067
  * :disable-na-as-missing for fixed types too
  * new `maximum` reducer for reductions namespace